+++ app/api/Datasets.scala
@@ -191,10 +191,10 @@ class  Datasets @Inject()(
       user match {
         case Some(identity) => {
           (request.body \ "space").asOpt[String] match {
-            case None | Some("default") => d = Dataset(name=name,description=description, created=new Date(), author=identity, licenseData = License.fromAppConfig())
+            case None | Some("default") => d = Dataset(name=name,description=description, created=new Date(), author=identity, licenseData = License.fromAppConfig(), stats = new Statistics())
             case Some(spaceId) =>
               spaces.get(UUID(spaceId)) match {
-                case Some(s) => d = Dataset(name=name,description=description, created=new Date(), author=identity, licenseData = License.fromAppConfig(), spaces = List(UUID(spaceId)))
+                case Some(s) => d = Dataset(name=name,description=description, created=new Date(), author=identity, licenseData = License.fromAppConfig(), spaces = List(UUID(spaceId)), stats = new Statistics())
                 case None => BadRequest(toJson("Bad space = " + spaceId))
               }
           }
@@ -274,7 +274,7 @@ class  Datasets @Inject()(
           case Some(identity) => {
             (request.body \ "space").asOpt[List[String]] match {
               case None | Some(List("default"))=> {
-                d = Dataset(name = name, description = description, created = new Date(), author = identity, licenseData = License.fromAppConfig(), status = access)
+                d = Dataset(name = name, description = description, created = new Date(), author = identity, licenseData = License.fromAppConfig(), stats = new Statistics(), status = access)
               }
 
               case Some(space) => {
@@ -287,7 +287,7 @@ class  Datasets @Inject()(
                     BadRequest(toJson("Bad space = " + aSpace))
                   }
                 }
-                d = Dataset(name = name, description = description, created = new Date(), author = identity, licenseData = License.fromAppConfig(), spaces = spaceList, status = access)
+                d = Dataset(name = name, description = description, created = new Date(), author = identity, licenseData = License.fromAppConfig(), spaces = spaceList, stats = new Statistics(), status = access)
               }
 
             }
@@ -2440,23 +2440,28 @@ class  Datasets @Inject()(
     Some(new ByteArrayInputStream(s.getBytes("UTF-8")))
   }
 
-  def download(id: UUID, compression: Int) = PermissionAction(Permission.DownloadFiles, Some(ResourceRef(ResourceRef.dataset, id))) { implicit request =>
+  def download(id: UUID, compression: Int, tracking: Boolean) = PermissionAction(Permission.DownloadFiles, Some(ResourceRef(ResourceRef.dataset, id))) { implicit request =>
     implicit val user = request.user
-        datasets.get(id) match {
-          case Some(dataset) => {
-            val bagit = play.api.Play.configuration.getBoolean("downloadDatasetBagit").getOrElse(true)
-            // Use custom enumerator to create the zip file on the fly
-            // Use a 1MB in memory byte array
-            Ok.chunked(enumeratorFromDataset(dataset,1024*1024, compression,bagit,user)).withHeaders(
-              CONTENT_TYPE -> "application/zip",
-              CONTENT_DISPOSITION -> (FileUtils.encodeAttachment(dataset.name+ ".zip", request.headers.get("user-agent").getOrElse("")))
-            )
-          }
-          // If the dataset wasn't found by ID
-          case None => {
-            NotFound
-          }
-        }
+    datasets.get(id) match {
+      case Some(dataset) => {
+        val bagit = play.api.Play.configuration.getBoolean("downloadDatasetBagit").getOrElse(true)
+
+        // Increment download count if tracking is enabled
+        if (tracking)
+          datasets.incrementDownloads(id, user)
+
+        // Use custom enumerator to create the zip file on the fly
+        // Use a 1MB in memory byte array
+        Ok.chunked(enumeratorFromDataset(dataset,1024*1024, compression,bagit,user)).withHeaders(
+          CONTENT_TYPE -> "application/zip",
+          CONTENT_DISPOSITION -> (FileUtils.encodeAttachment(dataset.name+ ".zip", request.headers.get("user-agent").getOrElse("")))
+        )
+      }
+      // If the dataset wasn't found by ID
+      case None => {
+        NotFound
+      }
+    }
   }
 
   def updateAccess(id:UUID, access:String) = PermissionAction(Permission.PublicDataset, Some(ResourceRef(ResourceRef.dataset, id))) { implicit request =>
@@ -2505,7 +2510,7 @@ class  Datasets @Inject()(
             if (u.id == dataset.author.id) {
               spaces.get(spaceId) match {
                 case Some(space) => {
-                  val d = Dataset(name = dataset.name, description = dataset.description, created = new Date(), author = dataset.author, licenseData = dataset.licenseData, spaces = List(spaceId))
+                  val d = Dataset(name = dataset.name, description = dataset.description, created = new Date(), author = dataset.author, licenseData = dataset.licenseData, spaces = List(spaceId), stats = dataset.stats)
                   datasets.insert(d) match {
                     case Some(id) => {
                       copyDatasetMetadata(dataset.id,UUID(id))
@@ -2524,7 +2529,7 @@ class  Datasets @Inject()(
                             val newFile = models.File(loader_id = file.loader_id, filename = file.filename, author = file.author,
                               uploadDate = file.uploadDate, contentType = file.contentType, length = file.length,
                               loader = file.loader, showPreviews = file.showPreviews,
-                              description = file.description, licenseData = file.licenseData, status = file.status)
+                              description = file.description, licenseData = file.licenseData, stats = file.stats, status = file.status)
                             files.save(newFile)
                             FileUtils.copyFileThumbnail(file,newFile)
                             FileUtils.copyFileMetadata(file,newFile)
@@ -2593,7 +2598,7 @@ class  Datasets @Inject()(
               val newFile = models.File(loader_id = file.loader_id, filename = file.filename, author = file.author,
                 uploadDate = file.uploadDate, contentType = file.contentType, length = file.length,
                 loader = file.loader, showPreviews = file.showPreviews, previews = file.previews, thumbnail_id = file.thumbnail_id,
-                description = file.description, licenseData = file.licenseData, status = file.status)
+                description = file.description, licenseData = file.licenseData, stats = file.stats, status = file.status)
               files.save(newFile)
               FileUtils.copyFileMetadata(file,newFile)
               FileUtils.copyFilePreviews(file,newFile)