+++ app/services/DatasetService.scala
@@ -1,5 +1,7 @@
 package services
 
+import java.util.Date
+
 import api.Permission.Permission
 import models._
 import play.api.libs.json.JsValue
@@ -377,6 +379,10 @@ trait DatasetService {
    * Move a creator to a new position in the dataset's list of creators
    */
   def moveCreator(id: UUID, creator: String, position: Integer)
+
+  def incrementViews(id: UUID, user: Option[User]): (Int, Date)
+
+  def incrementDownloads(id: UUID, user: Option[User])
 
 }