+++ app/services/mongodb/MongoDBDatasetService.scala
@@ -1514,7 +1514,7 @@ class MongoDBDatasetService @Inject() (
 		    return unsuccessfulDumps.toList
 	}
 
-    def dumpAllDatasetGroupings(): List[String] = {
+  def dumpAllDatasetGroupings(): List[String] = {
 
 		    Logger.debug("Dumping dataset groupings of all datasets.")
 
@@ -1588,6 +1588,45 @@ class MongoDBDatasetService @Inject() (
                     $pull("followers" -> new ObjectId(userId.stringify)), false, false, WriteConcern.Safe)
   }
 
+  def incrementViews(id: UUID, user: Option[User]): (Int, Date) = {
+    Logger.debug("updating views for dataset "+id.toString)
+    val viewdate = new Date
+
+    val updated = Dataset.dao.collection.findAndModify(
+      query=MongoDBObject("_id" -> new ObjectId(id.stringify)),
+      update=$inc("stats.views" -> 1) ++ $set("stats.last_viewed" -> viewdate),
+      upsert=true, fields=null, sort=null, remove=false, returnNew=true)
+
+    user match {
+      case Some(u) => {
+        Logger.debug("updating views for user "+u.toString)
+        DatasetStats.update(MongoDBObject("user_id" -> new ObjectId(u.id.stringify), "resource_id" -> new ObjectId(id.stringify), "resource_type" -> "dataset"),
+          $inc("views" -> 1) ++ $set("last_viewed" -> viewdate), true, false, WriteConcern.Safe)
+      }
+      case None => {}
+    }
+
+    // Return updated count
+    return updated match {
+      case Some(u) => (u.get("stats").asInstanceOf[BasicDBObject].get("views").asInstanceOf[Int], viewdate)
+      case None => (0, viewdate)
+    }
+  }
+
+  def incrementDownloads(id: UUID, user: Option[User]) = {
+    Logger.debug("updating downloads for dataset "+id.toString)
+    Dataset.update(MongoDBObject("_id" -> new ObjectId(id.stringify)),
+      $inc("stats.downloads" -> 1) ++ $set("stats.last_downloaded" -> new Date), true, false, WriteConcern.Safe)
+
+    user match {
+      case Some(u) => {
+        Logger.debug("updating downloads for user "+u.toString)
+        DatasetStats.update(MongoDBObject("user_id" -> new ObjectId(u.id.stringify), "resource_id" -> new ObjectId(id.stringify), "resource_type" -> "dataset"),
+          $inc("downloads" -> 1) ++ $set("last_downloaded" -> new Date), true, false, WriteConcern.Safe)
+      }
+      case None => {}
+    }
+  }
 }
 
 object Dataset extends ModelCompanion[Dataset, ObjectId] {
@@ -1618,3 +1657,9 @@ object LicenseData extends ModelCompanion[LicenseData, ObjectId] {
   }
 }
 
+object DatasetStats extends ModelCompanion[StatisticUser, ObjectId] {
+  val dao = current.plugin[MongoSalatPlugin] match {
+    case None => throw new RuntimeException("No MongoSalatPlugin");
+    case Some(x) => new SalatDAO[StatisticUser, ObjectId](collection = x.collection("statistics.users")) {}
+  }
+}